# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness Activity Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ActivitySummary {
  id: number;
  sport: string;
  startTime: string;
  /** @nullable */
  durationSeconds: number | null;
  /** @nullable */
  distanceMeters: number | null;
  /** @nullable */
  avgSpeedMps: number | null;
  /** @nullable */
  avgPaceSecPerKm: number | null;
  /** @nullable */
  totalElevGainMeters: number | null;
  /** @nullable */
  totalElevDescMeters: number | null;
  /** @nullable */
  maxSpeedMps: number | null;
  /** @nullable */
  avgHeartRate: number | null;
  /** @nullable */
  maxHeartRate: number | null;
  /** @nullable */
  totalCalories: number | null;
  /** @nullable */
  avgCadence: number | null;
  /** @nullable */
  avgPower: number | null;
  /** @nullable */
  normalizedPower: number | null;
  /** @nullable */
  avgVerticalOscillationMm: number | null;
  /** @nullable */
  avgStanceTimeMs: number | null;
  /** @nullable */
  avgVerticalRatio: number | null;
  /** @nullable */
  avgStepLengthMm: number | null;
  createdAt: string;
}

export interface DataPoint {
  timestamp: string;
  /** @nullable */
  heartRate: number | null;
  /** @nullable */
  cadence: number | null;
  /** @nullable */
  altitude: number | null;
  /** @nullable */
  lat: number | null;
  /** @nullable */
  lng: number | null;
  /** @nullable */
  speed: number | null;
  /** @nullable */
  distance: number | null;
  /** @nullable */
  power: number | null;
}

export interface ActivityDetail {
  id: number;
  sport: string;
  startTime: string;
  /** @nullable */
  durationSeconds: number | null;
  /** @nullable */
  distanceMeters: number | null;
  /** @nullable */
  avgSpeedMps: number | null;
  /** @nullable */
  avgPaceSecPerKm: number | null;
  /** @nullable */
  totalElevGainMeters: number | null;
  /** @nullable */
  totalElevDescMeters: number | null;
  /** @nullable */
  maxSpeedMps: number | null;
  /** @nullable */
  avgHeartRate: number | null;
  /** @nullable */
  maxHeartRate: number | null;
  /** @nullable */
  totalCalories: number | null;
  /** @nullable */
  avgCadence: number | null;
  /** @nullable */
  avgPower: number | null;
  /** @nullable */
  normalizedPower: number | null;
  /** @nullable */
  avgVerticalOscillationMm: number | null;
  /** @nullable */
  avgStanceTimeMs: number | null;
  /** @nullable */
  avgVerticalRatio: number | null;
  /** @nullable */
  avgStepLengthMm: number | null;
  /** @nullable */
  fileObjectPath: string | null;
  createdAt: string;
  dataPoints: DataPoint[];
}

export interface SportCount {
  sport: string;
  count: number;
}

export interface ActivityStats {
  totalActivities: number;
  totalDistanceMeters: number;
  totalDurationSeconds: number;
  avgDistanceMeters: number;
  avgDurationSeconds: number;
  sportBreakdown: SportCount[];
}

export interface StorageUploadRequestBody {
  name: string;
  size: number;
  contentType: string;
}

export type RenderJobStatus =
  (typeof RenderJobStatus)[keyof typeof RenderJobStatus];

export const RenderJobStatus = {
  queued: "queued",
  rendering: "rendering",
  complete: "complete",
  failed: "failed",
} as const;

export interface RenderJob {
  id: number;
  activityId: number;
  status: RenderJobStatus;
  progress: number;
  /** @nullable */
  videoObjectPath: string | null;
  /** @nullable */
  videoUrl: string | null;
  /** @nullable */
  errorMessage: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface StorageUploadPresignedUrl {
  uploadURL: string;
  objectPath: string;
}