# lib/api-zod/src/generated/types/renderJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness Activity Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { RenderJobStatus } from "./renderJobStatus";

export interface RenderJob {
  id: number;
  activityId: number;
  status: RenderJobStatus;
  progress: number;
  /** @nullable */
  videoObjectPath: string | null;
  /** @nullable */
  videoUrl: string | null;
  /** @nullable */
  errorMessage: string | null;
  createdAt: Date;
  updatedAt: Date;
}